User verification works correctly :)
