+++ src/utils/fetchUtils.js
@@ -1,7 +1,13 @@
+const fetchAuthToken = () => {
+  const authPayload = JSON.parse(localStorage.getItem('auth'))
+  return authPayload.accessToken
+}
+
 export const fetchConfig = () => {
   return {
     'Content-Type': 'application/json',
     'Accept': 'application/json',
+    'Authorization': `Bearer ${fetchAuthToken()}`
   };
 };
